feat/강좌API수정_#25

+++ src/main/java/com/seniorjob/seniorjobserver/controller/LectureStepTwoController.java
@@ -1,5 +1,7 @@
 package com.seniorjob.seniorjobserver.controller;
 
+import com.seniorjob.seniorjobserver.domain.entity.LectureEntity;
+import com.seniorjob.seniorjobserver.domain.entity.UserEntity;
 import com.seniorjob.seniorjobserver.domain.entity.WeekEntity;
 import com.seniorjob.seniorjobserver.dto.*;
 import com.seniorjob.seniorjobserver.repository.LectureRepository;
@@ -11,8 +13,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.parameters.P;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -43,8 +45,20 @@ public class LectureStepTwoController {
     @PostMapping("/{lectureId}/weeks")
     public ResponseEntity<WeekDto> addWeekTitle(
             @PathVariable Long lectureId,
-            @RequestParam String title
+            @RequestParam String title,
+            @AuthenticationPrincipal UserDetails userDetails
     ) {
+
+        UserEntity currentUser = userRepository.findByPhoneNumber(userDetails.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException("유저를 찾을 수 없습니다."));
+
+        // 현재 사용자가 강좌의 생성자인지 확인
+        LectureEntity lectureEntity = lectureRepository.findById(lectureId)
+                .orElseThrow(() -> new RuntimeException("강좌아이디 찾지못함 create_id: " + lectureId));
+        if (!lectureEntity.getUser().equals(currentUser)) {
+            throw new RuntimeException("해당 강좌의 주차별 제목을 작성할 권한이 없습니다.");
+        }
+
         WeekEntity newWeek = lectureStepTwoService.addWeekTitle(lectureId, title);
         WeekDto responseDto = new WeekDto(newWeek);
         return ResponseEntity.ok(responseDto);
@@ -55,8 +69,19 @@ public class LectureStepTwoController {
     public ResponseEntity<ExtendWeekPlanDto> addWeekPlan(
             @PathVariable Long lectureId,
             @PathVariable Long weekId,
-            @RequestParam String detail
+            @RequestParam String detail,
+            @AuthenticationPrincipal UserDetails userDetails
     ) {
+        UserEntity currentUser = userRepository.findByPhoneNumber(userDetails.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException("유저를 찾을 수 없습니다."));
+
+        // 현재 사용자가 강좌의 생성자인지 확인
+        LectureEntity lectureEntity = lectureRepository.findById(lectureId)
+                .orElseThrow(() -> new RuntimeException("강좌아이디 찾지못함 create_id: " + lectureId));
+        if (!lectureEntity.getUser().equals(currentUser)) {
+            throw new RuntimeException("해당 강좌의 제목의 주차별 상세내용을 작성할 권한이 없습니다.");
+        }
+
         ExtendWeekPlanDto newPlanDto = lectureStepTwoService.addWeekPlan(lectureId, weekId, detail);
         return ResponseEntity.ok(newPlanDto);
     }
@@ -65,7 +90,19 @@ public class LectureStepTwoController {
     @PostMapping("/{lectureId}/attendance")
     public ResponseEntity<AttendanceDto> setAttendanceCondition(
             @PathVariable Long lectureId,
-            @RequestParam int requiredAttendance){
+            @RequestParam int requiredAttendance,
+            @AuthenticationPrincipal UserDetails userDetails){
+
+        UserEntity currentUser = userRepository.findByPhoneNumber(userDetails.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException("유저를 찾을 수 없습니다."));
+
+        // 현재 사용자가 강좌의 생성자인지 확인
+        LectureEntity lectureEntity = lectureRepository.findById(lectureId)
+                .orElseThrow(() -> new RuntimeException("강좌아이디 찾지못함 create_id: " + lectureId));
+
+        if (!lectureEntity.getUser().equals(currentUser)) {
+            throw new RuntimeException("해당 강좌의 수료조건을 작성할 권한이 없습니다.");
+        }
 
         AttendanceDto attendanceDto = lectureStepTwoService.setAttendanceCondition(lectureId, requiredAttendance);
         return ResponseEntity.ok(attendanceDto);
@@ -88,6 +125,17 @@ public class LectureStepTwoController {
             @PathVariable Long weekId,
             @AuthenticationPrincipal UserDetails userDetails,
             @RequestParam String title){
+
+        UserEntity currentUser = userRepository.findByPhoneNumber(userDetails.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException("유저를 찾을 수 없습니다."));
+
+        // 현재 사용자가 강좌의 생성자인지 확인
+        LectureEntity lectureEntity = lectureRepository.findById(lectureId)
+                .orElseThrow(() -> new RuntimeException("강좌아이디 찾지못함 create_id: " + lectureId));
+        if (!lectureEntity.getUser().equals(currentUser)) {
+            throw new RuntimeException("해당 강좌의 주차별 제목을 수정할 권한이 없습니다.");
+        }
+
         WeekDto updatedWeek = lectureStepTwoService.updateWeekTitle(lectureId, weekId, title, userDetails);
         return ResponseEntity.ok(updatedWeek);
     }
@@ -99,6 +147,17 @@ public class LectureStepTwoController {
             @PathVariable Long lectureId,
             @PathVariable Long weekId,
             @AuthenticationPrincipal UserDetails userDetails){
+
+        UserEntity currentUser = userRepository.findByPhoneNumber(userDetails.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException("유저를 찾을 수 없습니다."));
+
+        // 현재 사용자가 강좌의 생성자인지 확인
+        LectureEntity lectureEntity = lectureRepository.findById(lectureId)
+                .orElseThrow(() -> new RuntimeException("강좌아이디 찾지못함 create_id: " + lectureId));
+        if (!lectureEntity.getUser().equals(currentUser)) {
+            throw new RuntimeException("해당 강좌의 주차별 제목을 삭제할 권한이 없습니다.");
+        }
+
         lectureStepTwoService.deleteWeek(lectureId, weekId, userDetails);
         return ResponseEntity.ok(String.format("%d강좌의 %d주차별제목과 해당 주차별 제목을 참조하는 " +
                 "주차별 상세내용들이 삭제되었습니다.", lectureId, weekId));
@@ -110,7 +169,18 @@ public class LectureStepTwoController {
             @PathVariable Long lectureId,
             @PathVariable Long weekId,
             @PathVariable Long planId,
-            @RequestParam String detail){
+            @RequestParam String detail,
+            @AuthenticationPrincipal UserDetails userDetails){
+
+        UserEntity currentUser = userRepository.findByPhoneNumber(userDetails.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException("유저를 찾을 수 없습니다."));
+
+        // 현재 사용자가 강좌의 생성자인지 확인
+        LectureEntity lectureEntity = lectureRepository.findById(lectureId)
+                .orElseThrow(() -> new RuntimeException("강좌아이디 찾지못함 create_id: " + lectureId));
+        if (!lectureEntity.getUser().equals(currentUser)) {
+            throw new RuntimeException("해당 강좌 -> 주차별제목의 상세내용을 수정할 권한이 없습니다.");
+        }
 
         WeekPlanDto updateWeekPlan = lectureStepTwoService.updateWeekPlan(lectureId, weekId, planId, detail);
         return ResponseEntity.ok(updateWeekPlan);
@@ -121,7 +191,19 @@ public class LectureStepTwoController {
     public ResponseEntity<String> deleteWeekPlan(
             @PathVariable Long lectureId,
             @PathVariable Long weekId,
-            @PathVariable Long planId){
+            @PathVariable Long planId,
+            @AuthenticationPrincipal UserDetails userDetails){
+
+        UserEntity currentUser = userRepository.findByPhoneNumber(userDetails.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException("유저를 찾을 수 없습니다."));
+
+        // 현재 사용자가 강좌의 생성자인지 확인
+        LectureEntity lectureEntity = lectureRepository.findById(lectureId)
+                .orElseThrow(() -> new RuntimeException("강좌아이디 찾지못함 create_id: " + lectureId));
+        if (!lectureEntity.getUser().equals(currentUser)) {
+            throw new RuntimeException("해당 강좌 -> 주차별제목의 상세내용을 삭제할 권한이 없습니다.");
+        }
+
         lectureStepTwoService.deleteWeekPlan(lectureId, weekId, planId);
         return ResponseEntity.ok(String.format("강좌번호 : %d, 주차제목번호 : %d, " +
                 "주차별상세내용 번호 %d 가 삭제되었습니다.", lectureId, weekId, planId));
@@ -131,12 +213,22 @@ public class LectureStepTwoController {
     @PutMapping("{lectureId}/update-attendance")
     public ResponseEntity<AttendanceDto> updateAttendanceCondition(
             @PathVariable Long lectureId,
-            @RequestParam int requiredAttendance) {
+            @RequestParam int requiredAttendance,
+            @AuthenticationPrincipal UserDetails userDetails) {
+
+        UserEntity currentUser = userRepository.findByPhoneNumber(userDetails.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException("유저를 찾을 수 없습니다."));
+
+        // 현재 사용자가 강좌의 생성자인지 확인
+        LectureEntity lectureEntity = lectureRepository.findById(lectureId)
+                .orElseThrow(() -> new RuntimeException("강좌아이디 찾지못함 create_id: " + lectureId));
+        if (!lectureEntity.getUser().equals(currentUser)) {
+            throw new RuntimeException("해당 강좌의 출석조건을 수정할 권한이 없습니다.");
+        }
 
         AttendanceDto updatedAttendance = lectureStepTwoService.updateAttendanceCondition(lectureId, requiredAttendance);
         return ResponseEntity.ok(updatedAttendance);
     }
-
 
     // 강좌개설 3단계 : 1,2단게에서 입력한 모든 정보를 확인한뒤 "강좌개설" 클릭시 강좌가 최종 개설되는 API Contorller
